StuttzPrototype_new/requirements.txt
flet==0.22.0
google-genai==1.0.0
python-dotenv==1.0.1
typing-extensions==4.10.0
